lcmsec: add circular test
added an integration test that tests group key agreement and subsequent message
passing within the group.

A total of 8 nodes send a message in a circle, this message includes a
way of addressing a specific node. An adressed increments the adress and rebroadcasts the message
Since all messages are multicast, one node (called client) can check if it has
sucessfully gotten a message by every other node. This shows that
encrypted group communication is possible.